Refactor virt-launcher.go into a separate package
